+++ example/lib/item_examples.dart
@@ -7,6 +7,7 @@ import 'src/actions_row.dart';
 import 'src/divider.dart';
 import 'src/item.dart';
 import 'src/position.dart';
+import 'src/pull_down_menu.dart';
 import 'src/selectable_item.dart';
 import 'src/show_menu.dart';
 import 'src/theming_custom.dart';
@@ -107,6 +108,10 @@ class ItemExamples extends StatelessWidget {
                   _Tile(
                     title: 'PullDownMenuItem.tapHandler',
                     destination: TapHandler(),
+                  ),
+                  _Tile(
+                    title: 'PullDownMenu',
+                    destination: PullDownMenuBox(),
                   )
                 ],
               ),