+++ lib/pull_down_button.dart
@@ -18,6 +18,7 @@ export 'src/items/entry.dart';
 export 'src/items/item.dart';
 export 'src/items/title.dart';
 export 'src/pull_down_button.dart';
+export 'src/pull_down_menu.dart';
 export 'src/theme/divider_theme.dart';
 export 'src/theme/item_theme.dart';
 export 'src/theme/route_theme.dart';